backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<3.15
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
PyJWT>=2.8,<3.0
pytest>=8.0
pytest-django>=4.8
